+++ hive/lib/src/annotations/generate_adapters.dart
@@ -1,3 +1,4 @@
+import 'package:hive_ce/src/annotations/hive_converter.dart';
 import 'package:meta/meta.dart';
 
 /// Annotation to generate TypeAdapters for the given [specs]
@@ -9,6 +10,7 @@ class GenerateAdapters {
     this.specs, {
     this.firstTypeId = 0,
     this.reservedTypeIds = const {},
+    this.converters = const [],
   });
   // coverage:ignore-end
 
@@ -22,6 +24,12 @@ class GenerateAdapters {
   ///
   /// These type ids will be skipped during generation
   final Set<int> reservedTypeIds;
+
+  /// A list of [HiveConverter]s to apply when generating adapters for [specs]
+  ///
+  /// These converters are used when a matching converter is not found on the
+  /// field or class. See [HiveConverter] for details.
+  final List<HiveConverter> converters;
 }
 
 /// Configuration that specifies the generation of a TypeAdapter
@@ -29,7 +37,10 @@ class GenerateAdapters {
 class AdapterSpec<T> {
   /// Constructor
   // coverage:ignore-start
-  const AdapterSpec({this.ignoredFields = const {}});
+  const AdapterSpec({
+    this.ignoredFields = const {},
+    this.converters = const [],
+  });
   // coverage:ignore-end
 
   /// Fields that should be ignored
@@ -37,4 +48,11 @@ class AdapterSpec<T> {
   /// This should only be used to simplify migrations from `HiveType`
   /// annotations. Model classes should only contain fields to be persisted.
   final Set<String> ignoredFields;
+
+  /// A list of [HiveConverter]s to apply when generating an adapter for [T]
+  ///
+  /// These converters take precedence over [GenerateAdapters.converters], but
+  /// are overridden by converters placed on the class or field. See
+  /// [HiveConverter] for details.
+  final List<HiveConverter> converters;
 }