+++ hive_generator/example/lib/hive/hive_adapters.dart
@@ -1,14 +1,30 @@
 import 'package:hive_ce/hive_ce.dart';
 import 'package:meta/meta.dart';
 
-@GenerateAdapters([
-  AdapterSpec<ClassSpec1>(),
-  AdapterSpec<ClassSpec2>(),
-  AdapterSpec<ClassSpec3>(),
-  AdapterSpec<ClassSpec4>(),
-  AdapterSpec<EnumSpec>(),
-], firstTypeId: 50)
+@GenerateAdapters(
+  [
+    AdapterSpec<ClassSpec1>(),
+    AdapterSpec<ClassSpec2>(),
+    AdapterSpec<ClassSpec3>(),
+    AdapterSpec<ClassSpec4>(),
+    AdapterSpec<EnumSpec>(),
+    AdapterSpec<ClassSpec5>(),
+  ],
+  firstTypeId: 50,
+  converters: [EpochDateTimeConverter()],
+)
 part 'hive_adapters.g.dart';
+
+/// Example converter matching json_serializable's JsonConverter pattern
+class EpochDateTimeConverter implements HiveConverter<DateTime, int> {
+  const EpochDateTimeConverter();
+
+  @override
+  DateTime fromHive(int hive) => DateTime.fromMillisecondsSinceEpoch(hive);
+
+  @override
+  int toHive(DateTime object) => object.millisecondsSinceEpoch;
+}
 
 @immutable
 class ClassSpec1 {
@@ -40,4 +56,12 @@ enum EnumSpec {
   value2;
 
   EnumSpec get getter => EnumSpec.value2;
+}
+
+@immutable
+class ClassSpec5 {
+  final DateTime time;
+  final DateTime? optionalTime;
+
+  const ClassSpec5(this.time, this.optionalTime);
 }