+++ hive_generator/lib/src/generator/adapters_generator.dart
@@ -96,6 +96,8 @@ class AdaptersGenerator extends GeneratorForAnnotation<GenerateAdapters> {
         typeId: schemaType.typeId,
         schema: schemaType,
         ignoredFields: spec.ignoredFields,
+        specConverters: spec.converters,
+        globalConverters: revived.converters,
       );
 
       content.write(result.content);